rabdan

+++ bot.py
@@ -4,8 +4,7 @@ import logging
 import config
 import asyncio
 import os
-import deepai
-import italygpt2
+import you
 
 from aiogram import Bot, Dispatcher, executor, types
 from aiogram.types.message import ContentType
@@ -13,7 +12,6 @@ from aiogram.utils import executor
 from aiogram.types import Message
 from aiogram.utils.deep_linking import get_start_link, decode_payload
 from aiogram.dispatcher import Dispatcher
-from aiogram.types import CallbackQuery
 
 import markup as nav
 from bot_db import Database
@@ -23,17 +21,20 @@ logging.basicConfig(level=logging.INFO)
 
 bot = Bot(token=config.TOKEN)
 dp = Dispatcher(bot)
+# database for local use
 db = Database('db/db.db')
+# database for docker container docker-compose up -d
+# db = Database('/root/db.db')
 
 async def setup_bot_commands(dp):
     bot_commands = [
         types.BotCommand("/newtopic", "Новая тема"),
         types.BotCommand("/about", "информация о боте"),
         types.BotCommand("/instruction", "Инструкция"),
-        types.BotCommand("/refferal", "Реферальная программа"),
+        types.BotCommand("/referral", "Реферальная программа"),
         types.BotCommand("/myprofile", "Профиль"),
         types.BotCommand("/subscribe", "Оплата подписки"),
-        ]
+    ]
     await dp.bot.set_my_commands(bot_commands)
 
 
@@ -72,7 +73,7 @@ async def start_message(message: types.Message) -> None:
             decrypted_link = Referr.decrypt_referral_link(message.text)
             referral_id = str(decrypted_link[7:])
 
-            await message.answer(f"Привет! Я – GPT-Dialog, языковая модель ИИ на базе ChatGPT, которая здесь, чтобы помочь. Просто спроси, что тебя интересует.") #reply_markup=nav.mainMenu)
+            await message.answer(config.TEXT_START)
 
             if str(referral_id) != '':
                 if int(referral_id) != int(message.from_user.id):
@@ -89,31 +90,14 @@ async def start_message(message: types.Message) -> None:
     
 @dp.message_handler(commands=("instruction"))
 async def instruction_info(message: types.Message) -> None:
-    await message.answer(f'''Работать со мной просто. Все, что вам нужно сделать, это ввести свои вопросы или подсказки, и я сгенерирую ответ на основе информации, на которой меня учили. Вот несколько советов для эффективного взаимодействия:
-- Используйте кнопку /newtopic для того, чтобы начать новый разговор, не относящийся контекстом к предыдущему.
-- Будьте ясны: формулируйте свои вопросы или подсказки четко и лаконично. Это помогает мне лучше понять ваши потребности и дать более точные ответы.
-- Предоставьте контекст: при необходимости предоставьте соответствующий контекст или справочную информацию, чтобы помочь мне лучше понять тему или ситуацию, о которой вы говорите. Чем больше подробностей вы предоставите, тем более индивидуальными и полезными могут быть мои ответы.
-- Задавайте уточняющие вопросы. Не стесняйтесь задавать уточняющие вопросы или запрашивать разъяснения, если это необходимо. Я здесь, чтобы помочь вам, поэтому не стесняйтесь вступать со мной в разговор.
-- Будьте вежливее: не забывайте поддерживать уважительный и вежливый тон во время нашего взаимодействия. Хотя я - искусственный интеллект, я ценю доброе и вежливое общение.
-- Есть ограничения в знаниях: хоть я и стараюсь предоставлять полезную и точную информацию, имейте в виду, что я не всегда могу иметь доступ к самой последней информации, и мои ответы не следует рассматривать как профессиональный совет.
-Помните, я здесь, чтобы помочь вам и предоставить информацию в меру своих возможностей. Просто введите ваши запросы, и я сделаю все возможное, чтобы помочь вам!''')#, reply_markup=nav.mainDMD)
+    await message.answer(config.TEXT_INSTRUCTION)
 @dp.message_handler(commands=("about"))
 async def give_info(message: types.Message) -> None:
-    await message.answer(f'''Как искусственный интеллект, я стараюсь помогать людям разными способами. Я могу предоставлять информацию по широкому кругу тем, отвечать на вопросы, предлагать предложения и участвовать в беседах. Если кому-то нужна помощь в решении сложной проблемы, он ищет совета по определенному вопросу или просто желает дружеского разговора, я здесь, чтобы протянуть руку помощи.
-Я стремлюсь предоставлять точную и полезную информацию, способствуя обучению и обмену знаниями. Однако важно отметить, что, хотя я могу генерировать ответы на основе шаблонов данных, на которых меня обучали, я не обладаю личным опытом, эмоциями или сознанием. Моя цель — давать полезные и информативные ответы на основе данных, на которых меня учили, но меня всегда следует использовать как инструмент для расширения человеческих возможностей, а не как замену профессионального совета или критического мышления.
-Таким образом, я являюсь языковой моделью ИИ, созданной для того, чтобы помогать и предоставлять информацию в меру своих возможностей. Моя цель — помогать людям, генерируя актуальные и ценные ответы, облегчая беседы и поддерживая пользователей в их начинаниях.''')#, reply_markup=nav.mainDMD)
-@dp.message_handler(commands=("refferal"))
+    await message.answer(config.TEXT_ABOUT)
+@dp.message_handler(commands=("referral"))
 async def give_info(message: types.Message) -> None:
-    await message.answer(f'''Реферальная программа: заработайте дополнительные дни бесплатного использования бота Telegram на основе подписки
-Мы рады представить нашу реферальную программу, предназначенную для вознаграждения наших уважаемых пользователей за распространение информации о нашем боте на основе подписки. Благодаря нашей реферальной программе у вас есть возможность продлить бесплатное использование бота, пригласив своих друзей, а также дав им дополнительный стимул присоединиться.
-Вот как работает реферальная программа:
-Период тестирования: когда вы впервые начнете использовать нашего бота, у вас будет один день, чтобы бесплатно изучить его возможности и функции. Это позволит вам ознакомиться с тем, что может предложить наш бот, прежде чем принять решение о подписке.
-Приглашение рефералов: после того, как вы попробуете бота и, если будете удовлетворены его возможностями, вы можете пригласить своих друзей и знакомых присоединиться. Поделитесь своей уникальной реферальной ссылкой в графе /Myprofile
-Бесплатное использование для вас: если ваш приглашенный друг решит подписаться и заплатит за месяц подписки бота, вы получите автоматическое продление на 5 дней бесплатного использования в дополнение к вашей существующей подписке. Это наш способ отблагодарить вас за рекомендацию нашего бота другим.
-Выгода для приглашенных рефералов: не только вы получаете выгоду от успешных рефералов, но и ваши приглашенные друзья! Они получат скидку на первый месяц использования в размере 10%!
-Наслаждайтесь постоянными преимуществами: по мере того, как вы набираете успешных рефералов, вы можете наслаждаться расширенным периодом бесплатного использования бота.
-Мы считаем, что наша реферальная программа обеспечивает беспроигрышную ситуацию как для вас, так и для ваших приглашенных друзей. Это позволяет вам продлить бесплатное использование бота, в то время как ваши друзья получают скидку при оплате.
-Спасибо за то, что являетесь частью нашего сообщества и помогаете нам расти. Мы ценим вашу поддержку и надеемся, что наш бот будет полезен и ценен в вашей повседневной деятельности. Если у вас есть дополнительные вопросы или вам нужна помощь, не стесняйтесь обращаться в нашу службу поддержки: ______________ (сюда пропишем юзера в Телеграм, надо создать аккаунт отдельный)''')#,reply_markup=nav.mainDMD )
+    await message.answer(config.TEXT_REFERRAL)
+
 
 @dp.message_handler(commands="newtopic")
 @dp.message_handler(lambda message: message.text and 'новая тема' in message.text.lower())
@@ -121,7 +105,7 @@ async def new_topic(message: types.Message) -> None:
     try:
         user_id = message.from_user.id
         messages[user_id] = []
-        messages[user_id].append({"role": "user", "content": "изначально ответы на русском языке"})
+        messages[user_id].append({"question": "изначально ответы на русском языке не более 150 слов", "answer": "конечно"})
 
         await message.answer('Начинаем новую тему!', parse_mode='Markdown')
     except Exception as e:
@@ -142,10 +126,11 @@ async def profile_handler(message: types.Message):
     encrypted_link = Referr.encrypt_referral_link(message.from_user.id)
     info = await bot.get_me()
     name = info.username
-    await message.answer(f'Дата окончания вашей подписки: {user_sub}\nВаша реферальная ссылка, чтобы поделиться с друзьями:\nhttps://telegram.me/{name}?start={encrypted_link}\nКоличество ваших рефералов: {db.count_referral(message.from_user.id)}')
+    await message.answer(f'Дата окончания: {date_string}\nID: {message.from_user.id}\nВаша реферальная сcылка:\nhttps://telegram.me/{name}?start={encrypted_link}\nКол-во рефералов:{db.count_referral(message.from_user.id)}')
+
 
 @dp.message_handler(commands="subscribe")
-@dp.message_handler(lambda message: message.text and 'submonth' in message.text.lower()) 
+@dp.message_handler(lambda message: message.text and 'оформить подписку' in message.text.lower())
 async def subscribe_handler(message: types.Message):
     try:
         if config.PAYMENTS_TOKEN.split(':')[1] == 'TEST':
@@ -181,6 +166,7 @@ async def subscribe_handler(message: types.Message):
     except Exception as e:
         logging.error(f'Error in subscribe: {e}')
 
+
 @dp.message_handler(content_types="text")
 async def send(message: types.Message):
     try:
@@ -194,18 +180,14 @@ async def send(message: types.Message):
             # asyncio.ensure_future(auto_delete_message(message.chat.id, processing_msg.message_id))
             if user_id not in messages:
                 messages[user_id] = []
-                messages[user_id].append({"role": "user", "content": "изначально ответы на русском языке"})
+                messages[user_id].append({"question": "изначально ответы на русском языке не более 150 слов", "answer": "конечно"})
 
-            messages[user_id].append({"role": "user", "content": user_message})
-            # messages[user_id].append({"role": "user", "content": f"chat: {message.chat} Сейчас {time.strftime('%d/%m/%Y %H:%M:%S')} user: {message.from_user.first_name} message: {message.text}"})
-            # Generate a response using OpenAI's Chat API
-            #account_data = italygpt2.Account.create()
-            #chatgpt_response_text = "".join(italygpt2.Completion.create(account_data=account_data,prompt="на русском языке",message=messages[user_id]))
-            chatgpt_response_text = "".join(deepai.ChatCompletion.create(messages=messages[user_id]))
+            chatgpt_response = you.Completion.create(prompt=user_message, chat=messages[user_id])
+            chatgpt_response_text = chatgpt_response.text.replace("\\/", "/").encode().decode('unicode_escape')
             # Add the bot's response to the user's message history
-            messages[user_id].append({"role": "assistant", "content": chatgpt_response_text})
+            messages[user_id].append({"question": user_message, "answer": chatgpt_response_text})
 
-            await message.answer(chatgpt_response_text) #reply_markup=nav.mainDMD)
+            await message.answer(chatgpt_response_text)
 
             await db.increment_counter_msg(user_id)
             await bot.delete_message(processing_msg.chat.id, processing_msg.message_id)
@@ -215,14 +197,16 @@ async def send(message: types.Message):
 
     except Exception as ex:
         # If an error occurs, try starting a new topic
-        await bot.delete_message(processing_msg.chat.id, processing_msg.message_id)
+        if 'processing_msg' in values():
+            await bot.delete_message(processing_msg.chat.id, processing_msg.message_id)
         print(ex)
         if ex == "context_length_exceeded":
-            await message.reply('У бота закончилась память, пересоздаю диалог',rparse_mode='Markdown')
+            await message.reply('У бота закончилась память, пересоздаю диалог', parse_mode='Markdown')
             await new_topic(message)
             await send(message)
         else:
-            await message.reply('У бота возникла ошибка, подождите некоторое время и попробуйте снова', parse_mode='Markdown')
+            await message.reply('У бота возникла ошибка, подождите некоторое время и попробуйте снова. Либо установите ограничение текста', parse_mode='Markdown')
+
 
 # pre checkout  (must be answered in 10 seconds)
 @dp.pre_checkout_query_handler(lambda query: True)
@@ -241,9 +225,8 @@ async def successful_payment(message: types.Message):
             referral_id = db.get_referral_id(message.from_user.id)
             print(referral_id)
             if referral_id is not None:
-                if (not db.get_date_status(message.from_user.id)):
-                    db.add_date_sub(referral_id, config.DAYS_ADD_FOR_REFFERAL)
-                    await bot.send_message(referral_id, f'Ваша подписка продлена на {config.DAYS_ADD_FOR_REFFERAL} дней, по Вашей реферальной ссылке зачислены средства')#reply_markup=nav.mainMenu)
+                db.add_date_sub(referral_id, config.DAYS_ADD_FOR_REFFERAL)
+                await bot.send_message(referral_id, f'Ваша подписка продлена на {config.DAYS_ADD_FOR_REFFERAL} дней, по Вашей реферальной ссылке зачислены средства')
 
     except Exception as e:
         logging.error(f'Error in payment: {e}')
